Modificari

+++ src/main/java/com/example/LibrarieSpring/service/ColectieService.java
@@ -2,8 +2,10 @@ package com.example.LibrarieSpring.service;
 
 import com.example.LibrarieSpring.controller.LoginController;
 import com.example.LibrarieSpring.entity.Colectie;
+import com.example.LibrarieSpring.entity.ElementColectie;
 import com.example.LibrarieSpring.entity.Utilizator;
 import com.example.LibrarieSpring.repository.ColectieRepository;
+import com.example.LibrarieSpring.repository.ElementColectieRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,7 +16,10 @@ import java.util.Optional;
 public class ColectieService {
 
     @Autowired
-    ColectieRepository cr;
+    private ColectieRepository cr;
+
+    @Autowired
+    private ElementColectieRepository ecolr;
 
     public boolean adaugaColectie(Colectie colectie)
     {
@@ -45,7 +50,18 @@ public class ColectieService {
 
     public void stergeColectieById(long id)
     {
-        cr.deleteById1(id);
+        Optional<Colectie> colectie=cr.findById(id);
+        if(colectie.isPresent()) {
+
+            List<ElementColectie> elementeColectie=ecolr.findByColectie(colectie.get());
+
+            for(ElementColectie ec:elementeColectie)
+            {
+                ecolr.deleteById1(ec.getId());
+            }
+
+            cr.deleteById1(id);
+        }
     }
 
     public Optional<Colectie> cautaColectieDupaId(long id)
